adding the Gitignore.txt
Add a .gitignore file to exclude any generated files such as binaries or
virtual environments. Any generated output must be excluded from
your Git repo.
For more information about creating a .gitignore file, we recommend
the following reading.
We have included a .gitignore file in this task file with common files
that should be added to a .gitignore file for a Django project.
You can use this file in the root folder of your project and add to it if
necessary. We suggest going through this file to see what types of files
are usually excluded from a project when pushing it to a remote
repository.
All of the packages and binary files that are required by the Django
project should be added again when somebody installs those
packages using pip